add project page and empty home page

+++ bazap-ui/src/Componenets/Pages/DeviceType/DeviceTypeTable.jsx
@@ -5,18 +5,9 @@ import Loader from "../../Layout/Loader";
 import { swalFire } from "../../UI/Swal";
 import { deleteDeviceType } from "../../../Utils/deviceTypeApi";
 import propTypes from "prop-types";
-import CustomModal from "../../UI/CustomModal";
-import { useState } from "react";
-import DeviceForm from "./DeviceForm";
 
 const DeviceTypeTable = ({ deviceTypes, isLoading }) => {
-    const [show, setShow] = useState(false);
-    const [formValues, setFormValues] = useState(null);
     const queryClient = useQueryClient();
-    const onEditDeviceTypeHandler = (data) => {
-        setFormValues({ deviceTypeName: data.deviceName, id: data._id });
-        showModal();
-    };
     const onDeleteDeviceTypeHandler = (id) => {
         swalFire({
             html: "האם אתה בטוח מעוניין למחוק את סוג המכשיר?",
@@ -32,7 +23,11 @@ const DeviceTypeTable = ({ deviceTypes, isLoading }) => {
         {
             name: "שם מכשיר",
             sortable: true,
-            selector: (row) => row.deviceName,
+            cell: (row) => {
+                const stringWithEntities = row.deviceName;
+                const decodedDeviceName = stringWithEntities.replace(/&#39;/g, "'");
+                return <div>{decodedDeviceName}</div>;
+            },
         },
         {
             name: "פעולות",
@@ -47,9 +42,6 @@ const DeviceTypeTable = ({ deviceTypes, isLoading }) => {
                     id="dropdown-item-button"
                     title="פעולות"
                 >
-                    <Dropdown.Item eventKey="edit" as="button">
-                        ערוך
-                    </Dropdown.Item>
                     <Dropdown.Item eventKey="delete" as="button" className="btn-light-danger">
                         מחק
                     </Dropdown.Item>
@@ -70,42 +62,17 @@ const DeviceTypeTable = ({ deviceTypes, isLoading }) => {
         },
     });
     const selectActionHandler = (data, event) => {
-        switch (event) {
-            case "edit":
-                onEditDeviceTypeHandler(data);
-                break;
-            case "delete":
-                onDeleteDeviceTypeHandler(data._id);
-                break;
-            default:
-                break;
+        if (event === "delete") {
+            onDeleteDeviceTypeHandler(data._id);
         }
     };
     if (isLoading) {
         return <Loader />;
     }
-    const showModal = () => {
-        setShow(true);
-    };
-    const hideModal = () => {
-        setShow(false);
-    };
-    const modalProperties = {
-        show,
-        handleClose: () => {},
-        title: "סוג מוצר חדש",
-        showExitButton: true,
-        showOkButton: false,
-        okButtonHandler: hideModal,
-        showCancelButton: false,
-        cancelButtonHandler: hideModal,
-    };
+
     return (
         <>
             <DataTable className="table" columns={columns} data={deviceTypes} />
-            <CustomModal {...modalProperties}>
-                <DeviceForm onCancel={hideModal} formValues={formValues} isEdit={true} />
-            </CustomModal>
         </>
     );
 };
